+++ src/components/index.ts
@@ -1,4 +1,5 @@
 import Navbar from './Navbar';
 import Footer from './Footer';
+import SocialLinks from './SocialLinks';
 
-export { Navbar, Footer };
+export { Navbar, Footer, SocialLinks };
